PM-21952: Move navigation package to UI module
