+++ app/src/main/kotlin/com/x8bit/bitwarden/ui/platform/components/model/ScaffoldNavigationData.kt
@@ -1,5 +1,6 @@
 package com.x8bit.bitwarden.ui.platform.components.model
 
+import com.bitwarden.ui.platform.components.navigation.model.NavigationItem
 import kotlinx.collections.immutable.ImmutableList
 
 /**
