+++ app/src/main/kotlin/com/x8bit/bitwarden/ui/platform/components/scaffold/BitwardenScaffold.kt
@@ -40,14 +40,14 @@ import androidx.compose.ui.semantics.testTagsAsResourceId
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.zIndex
 import com.bitwarden.ui.platform.base.util.toDp
+import com.bitwarden.ui.platform.components.navigation.BitwardenBottomAppBar
+import com.bitwarden.ui.platform.components.navigation.BitwardenNavigationRail
 import com.bitwarden.ui.platform.model.WindowSize
 import com.bitwarden.ui.platform.theme.BitwardenTheme
 import com.bitwarden.ui.platform.util.rememberWindowSize
 import com.x8bit.bitwarden.ui.platform.components.model.BitwardenPullToRefreshState
 import com.x8bit.bitwarden.ui.platform.components.model.ScaffoldNavigationData
 import com.x8bit.bitwarden.ui.platform.components.model.rememberBitwardenPullToRefreshState
-import com.x8bit.bitwarden.ui.platform.components.navigation.BitwardenBottomAppBar
-import com.x8bit.bitwarden.ui.platform.components.navigation.BitwardenNavigationRail
 import com.x8bit.bitwarden.ui.platform.components.scrim.BitwardenAnimatedScrim
 
 /**
