+++ ui/src/main/kotlin/com/bitwarden/ui/platform/components/navigation/color/BitwardenNavigationBarItemColors.kt
@@ -1,15 +1,15 @@
-package com.x8bit.bitwarden.ui.platform.components.navigation.color
+package com.bitwarden.ui.platform.components.navigation.color
 
-import androidx.compose.material3.NavigationRailItemColors
+import androidx.compose.material3.NavigationBarItemColors
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import com.bitwarden.ui.platform.theme.BitwardenTheme
 
 /**
- * Provides a default set of Bitwarden-styled colors for navigation rail items.
+ * Provides a default set of Bitwarden-styled colors for navigation bar items.
  */
 @Composable
-fun bitwardenNavigationRailItemColors(): NavigationRailItemColors = NavigationRailItemColors(
+fun bitwardenNavigationBarItemColors(): NavigationBarItemColors = NavigationBarItemColors(
     selectedIconColor = BitwardenTheme.colorScheme.icon.secondary,
     unselectedIconColor = BitwardenTheme.colorScheme.icon.primary,
     disabledIconColor = BitwardenTheme.colorScheme.outlineButton.foregroundDisabled,
